examples/text2sql/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 64", "setuptools-scm>=8"]

[tool.setuptools_scm]
# NAT uses the --first-parent flag to avoid tags from previous releases which have been merged into the develop branch
# from causing an unexpected version change. This can be safely removed if developing outside of the NAT repository.
git_describe_command = "git describe --long --first-parent"
root = "../.."

[project]
name = "nat-text2sql"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[langchain]~=1.3.0rc3",
  "vanna>=0.3.4",
  "pymilvus>=2.3.0",
  "databricks-sql-connector>=3.0.0",
  "sqlglot>=20.0.0",
  "pandas>=2.0.0",
  "python-dotenv>=1.0.0",
]
requires-python = ">=3.11,<3.14"
description = "Standalone Text2SQL example for MCP server deployment and load testing"
classifiers = ["Programming Language :: Python"]

[tool.uv.sources]
nvidia-nat = { path = "../..", editable = true }

[project.entry-points.'nat.components']
text2sql = "text2sql.register"
